+++ Package.swift
@@ -26,7 +26,7 @@ let package = Package(
       .revision("swift-DEVELOPMENT-SNAPSHOT-2020-01-29-a")
     ),
     .package(url: "https://github.com/apple/swift-tools-support-core.git", from: "0.0.1"),
-    .package(url: "https://github.com/apple/swift-argument-parser.git", .upToNextMinor(from: "0.0.2")),
+    .package(url: "https://github.com/apple/swift-argument-parser.git", .upToNextMinor(from: "0.0.4")),
   ],
   targets: [
     .target(
