+++ Sources/swift-format/Subcommands/Format.swift
@@ -2,7 +2,7 @@
 //
 // This source file is part of the Swift.org open source project
 //
-// Copyright (c) 2014 - 2019 Apple Inc. and the Swift project authors
+// Copyright (c) 2014 - 2020 Apple Inc. and the Swift project authors
 // Licensed under Apache License v2.0 with Runtime Library Exception
 //
 // See https://swift.org/LICENSE.txt for license information
@@ -10,59 +10,62 @@
 //
 //===----------------------------------------------------------------------===//
 
+import ArgumentParser
 import Foundation
 import SwiftFormat
 import SwiftFormatConfiguration
-import SwiftFormatCore
 import SwiftSyntax
 import TSCBasic
 
-/// Runs the linting pipeline over the provided source file.
-///
-/// If there were any lint diagnostics emitted, this function returns a non-zero exit code.
-/// - Parameters:
-///   - configuration: The `Configuration` that contains user-specific settings.
-///   - sourceFile: A file handle from which to read the source code to be linted.
-///   - assumingFilename: The filename of the source file, used in diagnostic output.
-///   - ignoreUnparsableFiles: Whether or not to ignore files that contain syntax errors.
-///   - debugOptions: The set containing any debug options that were supplied on the command line.
-///   - diagnosticEngine: A diagnostic collector that handles diagnostic messages.
-/// - Returns: Zero if there were no lint errors, otherwise a non-zero number.
-func lintMain(
-  configuration: Configuration, sourceFile: FileHandle, assumingFilename: String?,
-  ignoreUnparsableFiles: Bool, debugOptions: DebugOptions, diagnosticEngine: DiagnosticEngine
-) {
-  let linter = SwiftLinter(configuration: configuration, diagnosticEngine: diagnosticEngine)
-  linter.debugOptions = debugOptions
-  let assumingFileURL = URL(fileURLWithPath: assumingFilename ?? "<stdin>")
+extension SwiftFormatCommand {
+  /// Formats one or more files containing Swift code.
+  struct Format: ParsableCommand {
+    static var configuration = CommandConfiguration(
+      abstract: "Format Swift source code",
+      discussion: "When no files are specified, it expects the source from standard input.")
 
-  guard let source = readSource(from: sourceFile) else {
-    diagnosticEngine.diagnose(
-      Diagnostic.Message(.error, "Unable to read source for linting from \(assumingFileURL.path)."))
-    return
-  }
+    /// Whether or not to format the Swift file in-place.
+    ///
+    /// If specified, the current file is overwritten when formatting.
+    @Flag(
+      name: .shortAndLong,
+      help: "Overwrite the current file when formatting.")
+    var inPlace: Bool
 
-  do {
-    try linter.lint(source: source, assumingFileURL: assumingFileURL)
-  } catch SwiftFormatError.fileNotReadable {
-    let path = assumingFileURL.path
-    diagnosticEngine.diagnose(
-      Diagnostic.Message(.error, "Unable to lint \(path): file is not readable or does not exist."))
-    return
-  } catch SwiftFormatError.fileContainsInvalidSyntax(let position) {
-    guard !ignoreUnparsableFiles else {
-      return
+    @OptionGroup()
+    var formatOptions: LintFormatOptions
+
+    func validate() throws {
+      if inPlace && formatOptions.paths.isEmpty {
+        throw ValidationError("'--in-place' is only valid when formatting files")
+      }
     }
-    let path = assumingFileURL.path
-    let location = SourceLocationConverter(file: path, source: source).location(for: position)
-    diagnosticEngine.diagnose(
-      Diagnostic.Message(.error, "file contains invalid or unrecognized Swift syntax."),
-      location: location)
-    return
-  } catch {
-    let path = assumingFileURL.path
-    diagnosticEngine.diagnose(Diagnostic.Message(.error, "Unable to lint \(path): \(error)"))
-    return
+
+    func run() throws {
+      let diagnosticEngine = makeDiagnosticEngine()
+
+      if formatOptions.paths.isEmpty {
+        let configuration = try loadConfiguration(
+          forSwiftFile: nil, configFilePath: formatOptions.configurationPath)
+        formatMain(
+          configuration: configuration, sourceFile: FileHandle.standardInput,
+          assumingFilename: formatOptions.assumeFilename, inPlace: false,
+          ignoreUnparsableFiles: formatOptions.ignoreUnparsableFiles,
+          debugOptions: formatOptions.debugOptions, diagnosticEngine: diagnosticEngine)
+      } else {
+        try processSources(
+          from: formatOptions.paths, configurationPath: formatOptions.configurationPath,
+          diagnosticEngine: diagnosticEngine
+        ) { sourceFile, path, configuration in
+          formatMain(
+            configuration: configuration, sourceFile: sourceFile, assumingFilename: path,
+            inPlace: inPlace, ignoreUnparsableFiles: formatOptions.ignoreUnparsableFiles,
+            debugOptions: formatOptions.debugOptions, diagnosticEngine: diagnosticEngine)
+        }
+      }
+
+      try failIfDiagnosticsEmitted(diagnosticEngine)
+    }
   }
 }
 
@@ -77,7 +80,7 @@ func lintMain(
 ///   - debugOptions: The set containing any debug options that were supplied on the command line.
 ///   - diagnosticEngine: A diagnostic collector that handles diagnostic messages.
 /// - Returns: Zero if there were no format errors, otherwise a non-zero number.
-func formatMain(
+private func formatMain(
   configuration: Configuration, sourceFile: FileHandle, assumingFilename: String?, inPlace: Bool,
   ignoreUnparsableFiles: Bool, debugOptions: DebugOptions, diagnosticEngine: DiagnosticEngine
 ) {
@@ -136,11 +139,4 @@ func formatMain(
     diagnosticEngine.diagnose(Diagnostic.Message(.error, "Unable to format \(path): \(error)"))
     return
   }
-}
-
-/// Reads from the given file handle until EOF is reached, then returns the contents as a UTF8
-/// encoded string.
-fileprivate func readSource(from fileHandle: FileHandle) -> String? {
-  let sourceData = fileHandle.readDataToEndOfFile()
-  return String(data: sourceData, encoding: .utf8)
 }
