+++ Sources/swift-format/Utilities/FileIterator.swift
